+++ apps/web/src/routes/home.tsx
@@ -14,7 +14,8 @@ import { Window } from "../shell/window";
 import { SettingsWindow } from "../shell/settings-window";
 import { HistoryWindow } from "../shell/history-window";
 import { AddressWindow } from "../shell/address-window";
-import { EmbedWindow } from "../shell/embed-window";
+import { ExplorerWindow } from "../shell/explorer-window";
+import { GnockpitEmbedWindow } from "../shell/gnockpit-embed-window";
 import { ExtraRealmWindows } from "../shell/extra-realm-windows";
 import { useWindowPersistence } from "../shell/use-window-persistence";
 import { useWindowViewportReclamp } from "../shell/use-window-viewport-reclamp";
@@ -161,7 +162,8 @@ export function Home() {
           <SettingsWindow />
           <HistoryWindow />
           <AddressWindow />
-          <EmbedWindow />
+          <ExplorerWindow />
+          <GnockpitEmbedWindow />
           <ExtraRealmWindows />
         </div>
       </div>